[utils] Move json_writer from daemon to utils.
