+++ src/daemon/default_vm_image_vault.cpp
@@ -16,11 +16,11 @@
  */
 
 #include "default_vm_image_vault.h"
-#include "json_writer.h"
 
 #include <multipass/exceptions/aborted_download_exception.h>
 #include <multipass/exceptions/create_image_exception.h>
 #include <multipass/exceptions/unsupported_image_exception.h>
+#include <multipass/json_writer.h>
 #include <multipass/logging/log.h>
 #include <multipass/platform.h>
 #include <multipass/process/qemuimg_process_spec.h>
